Grab the html version of console log
Currently we are only grabbing the text output from Jenkins. However,
this doesn't have timestamps which isn't ideal. We can use the
internal url that is used by Jenkins to generate its web page to
fetch the console logs with timestamps and formatting.

We need to surround the output in pre tags since os-loganlayze uses
that to determine whether to escape the html or not.

Change-Id: I7584ea37f84521d65f327f327ff5e47b096fa230