Add test jobs for devstack-precise-dib images
Next step in migrating to dib is moving to devstack-precise-dib images
in hpcloud. We have images and slaves built with dib for this label now
we need to run tests on those nodes. Add experimental jobs to
devstack-gate for nova network and neutron run against icehouse on these
new -dib images.

Change-Id: I3efb17d93173091827623c15704e5a79be8fca33