Grab the html version of console log
Currently we are only grabbing the text output from Jenkins. However,
this doesn't have timestamps which isn't ideal. We can use the
internal url that is used by Jenkins to generate its web page to
fetch the console logs with timestamps and formatting.

This is the update for the jenkins macros.
Co-Authored-By: Joshua Hesketh <josh@nitrotech.org>

Change-Id: I1a092514cbc4e29d783dd120c69cc15f32885b48