Promote ironic sideways grenade to voting
The job is currently voting in the places where it runs today.  This
ensures the existing jobs are named appropriately (removes -nv) and adds
it to the gate of the places its running in check.  It also adds it to nova
check/gate pipelines.  The job will only ever run against juno branches,
so it is removed from devstack-gate which does not have stable branches.

Change-Id: Ied48252487f72517a1e8361f90ef49ae8d454885