Removed python26 from python-jobs
This patch removes the python26 jobs from the default python-jobs
template, and adds them manually to each project that was previously
using them. This patch is not intended to deprecate, pin, or remove
the python26 build from any project; instead, it is intended to support
such activity that may occur in subsequent patches.

Change-Id: I3124e9e557d7e537d173a4f6c421e55b17e693da