Use a job-template for the tripleo jobs
Now that we are setting the tripleo control variables in tripleo-ci we
can use a template for all of the tripleo jobs as the only difference
between them is the job name.

This will slightly increase the timeout for the undercloud jobs but the
impact of this side effect should be minimal.

Change-Id: I532d79e260e44e95d72e2372b1afb801c3bddf4d