Don't run gate-{name}-pypy for stable/icehouse
* zuul/layout.yaml: Since we're only supporting PyPy on Ubuntu
Trusty but we're not supporting stable/icehouse on Ubuntu Trusty,
we're transitively also not supporting PyPy for stable/icehouse.
Based on a review of the zuul layout, the only project I can find
with a stable/icehouse branch and gate-{name}-pypy jobs from the
pypy-jobs project template or otherwise is openstack/requirements.

Change-Id: Ia109de2ffeb242847f964fb9b636572d84c4a2fd