Allow non standard docker-py in check-tempest-dsvm-f20-docker
We are investigating the use of docker-py to get rid of a custom
http client and switch to a standard library docker-maintained
by docker folks. So we need these dsvm jobs at the very least
to see what works and what doesn't and how much regression we have

There are 2 dsvm jobs for nova-docker, we are making the change
of switching on NON_STANDARD_REQS only for the f20 one. As we
just need data points to make sure we are not regressing the
nova-docker driver as compared to existing functionality in the
review mentioned below. Best case scenario, switching on the
flag works as advertised and the requirements check will not
fail the dsvm job at the point it is failing now. Worst case
scenario the flag has no effect and we can delete the line
flag at that point.

Review against nova-docker:
I3e17924b8319720ceb7cee480a81e953b809d0ee

Review for docker-py in global requirements:
I3100ad32a1cb414033a3aee495ddd6b82e6af268

Change-Id: I1fb4d58a36c9a12f87527f4fb551f4b664c386bd