remove non-standard-reqs flag from check-tempest-dsvm-f20-docker
Introduced a flag in I1fb4d58a36c9a12f87527f4fb551f4b664c386bd, this
does not work when we set it from here. So this review removes the
NON_STANDARD_REQS flag. The reason it does not work is that
devstack runs stack.sh with sudo and the environment variable is not
passed down into the process.

Change-Id: I379a5e6c566e8ac207c70089f3251eff6afe3379