Add openstack-infra/puppet-httpd
We currently pin to puppetlabs-apache 0.0.4 (released 09-05-2012).

The main difference with later versions is that they stopped
supporting using a template for vhost configuration; rather moving
various parts of configuratoin as option arguments.

This means that to convert to any later version of the module it
requires going through all of our templates; currently something like:

---
[system-config]$ find | grep vhost.erb
./modules/logstash/templates/logstash.vhost.erb
./modules/logstash/templates/kibana.vhost.erb
./modules/openstack_project/templates/puppetboard/puppetboard.vhost.erb
./modules/openstack_project/templates/logs.vhost.erb
./modules/openstack_project/templates/status.vhost.erb
./modules/openstack_project/templates/logs-dev.vhost.erb
./modules/drupal/templates/drupal.vhost.erb
./modules/gerrit/templates/gerrit.vhost.erb
./modules/meetbot/templates/vhost.erb
./modules/openstackid/templates/vhost.erb
./modules/jenkins/templates/jenkins.vhost.erb
./modules/cgit/templates/git.vhost.erb
./modules/zuul/templates/zuul.vhost.erb
./modules/etherpad_lite/templates/etherpadlite.vhost.erb
./modules/mailman/templates/mailman.vhost.erb
./modules/graphite/templates/graphite.vhost.erb
---

figuring out what they do, figuring out what arguments to translate
this to and redeploying all of these without breaking anything.

Even if we want to upgrade to the latest versions, which is a matter
of some debate, doing it all at once is almost impossible.

This creates openstack-infra/puppet-httpd which will be a fork of
0.0.4.  This can cover us both ways -- we can upgrade piecemeal, or
stay with the status-quo and add small, appropriate enhancements to
this module without having to upgrade the whole world.

The name was chosen by committee [1] to avoid conflict with the
original module (should we wish to use that in parallel) and to avoid
any confusion that this might be used to deploy apache for OpenStack
projects in general.

[1] http://eavesdrop.openstack.org/meetings/infra/2014/infra.2014-11-11-19.00.html

Change-Id: If6354fa3f4123fb4441a349881b17f87f6d8934b