Only run stackforge/puppet-* unit tests on trusty
Without this patch, stable/icehouse branches in stackforge/puppet-*
will run unit tests on Ubuntu Precise. This is a problem because the
version of Ruby and therefore Facter is too old on Precise and causes
failures for certain Facter values only available in Facter > 2. This
patch removes precise from the nodes to test the puppet-* modules on
and adds the puppet unit tests as an exception to set_node_options() to
prevent zuul from trying to use precise on stable/icehouse.

Change-Id: Id7453171a8ca2321528e0e0d3d2a0984e0304db3