Add a new qa-specs-core group and use it
This commit changes the core group on the qa-specs repo to be a
separate group. When qa-specs started the only project it affected was
tempest, however for kilo we are going to be using it for devstack and
grenade too. To manage domain expertise in the separate projects we need
to be able to hand tailor the core group to include the appropriate sets
of people which we couldn't do if it was just using the core group from
tempest. So a new group is needed to handle this.

Change-Id: I3e3f0da0f2aea5903017444167c472867885a753