Remove icehouse heat slow job from tempest check
This commit removes the stable heat slow job from the tempest check
queue. This jobs were added originally to ensure branchless tempest
continued to work against the stable branches in all gating configs.
But, because of some heat bugs and changes in testing methodology this
job doesn't add much value for icehouse. That coupled with the recent
nodepool image cache issues we've been hitting it makes more sense to
just remove the job. If the value of the job is increased and after
nodepool issues are sorted we can add this back.

Change-Id: I88b61b6aa70a1e336aeff66560134d7444f91880