JSHint job updates
The JSHint job was added during Juno and the style-changes-only patches
that were written to make it pass are unlikely to be backported to the
previous stable branches. Stop running it against stable/icehouse.

Now that the job is passing both on master and stable/juno, it would
make sense to make it voting to keep things that way.

Finally, make sure to update the apt-get cache before installing npm
in order to be more future-proof against future package updates.

Change-Id: I0ff4bad8af7507d039390358bd43df1348ed2e98