Use full path to grub2-mkconfig
Current centos7 builds on hpcloud are failing calling grub2-mkconfig;
from the logs:

 + sudo grub2-mkconfig -o /boot/grub2/grub.cfg
 sudo: grub2-mkconfig: command not found

To be completely honest, I'm having a hard time explaining why this
works with rackspace images, but not hpcloud ones.  I can not
replicate this in my own setup, even using the same hpcloud images.
Possibly it's something to do with the nodepool host environment when
it starts the script on the remote host?  Without access to a failing
host I'm clutching at straws...

Change-Id: I9d9f759fac1a96cc785354ae5aa44398c908704d