change client libraries to test from pypi instead of git
We are making devstack use client libraries from pypi instead of from
git. Before we do that we need to change the client library job
definitions to support testing themselves against changes in their own
git trees. This is equivalent to what was done with the oslo libraries.

This adds the lib-forward-testing to the client libraries for testing,
and purges the integrated gate jobs from them. It also changes the
stable-compat jobs to be driving by -src-{name} jobs, which will test
a client library with openstack servers from git, all the rest of the
client libraries from pypi, and the {name} library from git. So a
python-novaclient change will be tested with nova stable/*, released
glance/cinder clients, but the new zuul ref for python-novaclient.

Co-Authored-By: Joshua Hesketh <josh@nitrotech.org>

Needed-By: I26fac0ccf8fd4818e24618d56bf04b32306f88f6

Change-Id: I81b0d228e7769758c61e5b0323ecfce8c8886d39