Move zuul_swift_upload to publisher
Failed tests aren't being uploaded to swift due to the build process
exiting before the swift upload script is called. To fix this, run
the upload process as a publisher script.

We'll re-purpose the existing experimental job to test it. Thus make
the job always fail so we can then check the logs are uploaded.

Depends-on: I5ff21f544e6f2aa4bf88b7d02e65cdfb62e6faaf

Change-Id: I3f15604d8c3c2f405dc37a9d0bd4b044228e5793