Change the periodic tempest "all" job to use neutron
The slow heat job is failing on the nightly tempest 'all' run in the
CfnInitScenarioTest.test_server_cfn_init test, but passes in the
slow heat job. The main differences seem to be that the nightly job runs a lot
more stuff (more stress), and the nightly job uses nova-network but the slow
heat job uses neutron. Changing the 'all' job to neutron will help debug
this and at this point we should be using neutron anyway.

Change-Id: I74afab102c1dcf52aef15b0f4f4f15bbfc08250a