remove the dsvm-cells job from libraries
Libraries no longer test in a way that these jobs provide any
meaningful results. We should remove them from all the libraries. If
there are specific pieces of function that need testing in these
libraries, we should get them back into the libraries themselves.

This should also remove the coupling that oslo was still seeing with
the rest of the integrated gate.

Change-Id: I889630201e238a300fb16a39f8e43f5d6f3f635e