Make agent_ssh check job voting for IPA
The agent_ssh-src job has been passing for a week and should be reliable
enough to vote. We'd like this job to vote on IPA changes to prevent
breaking changes from merging.

I've also removed the buildimage-coreos jobs from the check/gate for
IPA, as the agent_ssh-src job will not pass unless it is able to
successfully build a working IPA image. However, it remains in post so
it can continue publishing to tarballs.o.o.

As a note; this does mean that Ironic can break IPA's gate (because
Ironic isn't voting on agent tempest jobs), but this is a small risk as
the core teams between the projects overlap significantly.

Change-Id: Idd3cf25612aaed4f1c28bfaba94bba413827b378