Add a project for OpenStack meetings descriptions
openstack-infra/gerrit-powered-agenda is being split into two
repositories: one for the code and one for the meetings descriptions
(so that review teams can be different and the code part can be reused).

This commit adds the repository for the meeting descriptions. After a
discussion with jeblair, I renamed it openstack-infra/irc-meetings.

Change-Id: Id75091087b7e36a17476fc6f47dad6d5ba473624