Add job to publish puppet modules to the forge
Puppet blacksmith is a ruby gem that can handle pushing new puppet
modules to the puppet forge. Right now we don't have an automated
way to push to the forge, so this is a big win. Added it to the
'release' pipeline. This job runs on the 'pypi' node which is
pypi.slave.o.o, this node is not a single use node and is trusted
with secrets for pypi uploads, maven uploads, and jenkins plugin
uploads. When this node goes trusty, it might make sense for it
to be renamed 'trusted.'

A limitation of puppet-blacksmith is that it can only source one
set of credentials. I am working on patches to allow it to handle
different credentials with a config option. This would be needed
to publish both the openstack-infra modules and the stackforge
modules.

Puppet-blacksmith does not run any user-submitted code, so it is
safe to do the entire operation on the trusted node.

Change-Id: I0c122828df6f231d1c20b8680e5bf2c90894ad62