Filter out some Tempest tests for Nova cells
We would like to get a set of Tempest tests running against Nova cells
that is expected to pass given the current functionality provided.
This is being done via regexes so that we can spend the bulk of our
effort on cellsv2 and put the current cells feature into maintenance
mode quickly.

Change-Id: I0c6047ca66d4d740f8ca0bc67a258bb36db8dbfe