remove python-heatclient from integrated gate
apparently when cleaning up the python-*clients to not be in the
integrated gate, I missed removing that from heatclient. In the libs
from pypi world the integrated gate jobs don't do what you think they
do, and the source jobs are really the important ones.

Will also let python-heatclient merge code faster as it will be
decoupled from the integrated gate pipeline.

Change-Id: I45bb269d700cbd9cdf90b6b9d99df18ded444fdf