Initial magnum client import for stackforge
Magnum needs a pythonclient in order to be usable from the CLI or
pyhthon applications.  While the initial upstream version does nothing
but pass the gate, we intend to grow this to provide the following
API:

* bay list
* bay create
* bay delete
* bay show

* pod create
* pod list
* pod delete
* pod show

* service create
* service list
* service delete
* service show

* container create
* container list
* container delete
* container show
* container reboot
* container stop
* container start
* container pause
* container unpause
* container logs
* container execute

The magnum core team feels a python client is mandatory and kindly
ask for the creation of a repository.

Change-Id: Ifedbf4799b914a24e139e4e6a416134aada7b623