Add stable compat jobs to oslo libs in stable requirements
This commit plugs a testing gap which we've hit at least twice in the
last 2 days. Oslo libraries that are in stable branches of global
requirements uncapped can add changes which will break dsvm jobs on
the stable branches. This adds the stable compat checking to those
jobs to ensure we don't break the stable branches with new changes.

Change-Id: Ib0ebac021afadb9d829a2adaadbacf76aea5fe55