Add Networking-vSphere project to StackForge
This repo will hold the Neutron drivers and agents
necessary to manage vSphere management clusters as
proposed by specifications [1, 2]. Other fully open
source contributions to Neutron support for vSphere
are also welcome, and this is just a start.

The need of setting this repo up is inline with
spec proposal [3], where core elements of the Neutron
project are decoupled by the internals of the
technology being supported.

[1] https://review.openstack.org/#/c/103728/
[2] https://review.openstack.org/#/c/104452/
[3] https://review.openstack.org/#/c/134680/

Change-Id: I0c883e0ab0e0d97f8f11da008199cae43bf3cb00