Add Castellan project
Castellan is a Key Manager interface for OpenStack projects that do not
want to integrate with Barbican directly.  Barbican will provide an
implementation of this interface, but projects can choose to implement
the interface with an alternative Key Management System if Barbican does
not meet their requirements.

A CR has laso been sent to the governance repo to add Castellan to the
Barbican program:

https://review.openstack.org/#/c/138875/

Change-Id: I076e04e5167e2ff00b94142e86c242de36e766da