only test libraries in a neutron config
The library testing jobs aren't really about testing the libraries,
they are about sniff testing them in a live deploy. By only using the
neutron jobs we are effectively testing the entirety of the integrated
release, minus a *very* small amount of code in nova.

This will reduce the amount of random fail on these jobs, use a lot
less nodes, and achieve effectively the same results. Any n-net only
breaks that slip through should be addressed by other mechanisms,
because the likelyhood that n-net is so dramatically using libraries
differently from the rest of nova is vanishlying small.

Change-Id: I1dcd8558b98443ffa9a3a01c2e877bd64bb96296