remove nova pylint
This has been broken for weeks because of pylint upgrade changing
format. Pylint is becoming less and less useful as it doesn't do sane
things when presented with nova objects, so burning a test node on
every proposed patch doesn't seem incredibly useful if most of the
time it's not useful at all.

Change-Id: I4defa01608f651157120faf5a8eff70f74203baa