Added gate-{name}-chef-rake job-template
This job template is supposed to be the first try of implementing rake
for managing the tests and how they are run. Added installation of rake gem to
ruby-191-prep macro to make it available for chef-rake-test macro and
skip the usage of chef-bundler-prep (done in rake bundle_prep in
chef-rake-test). Added a new project-template chef-rake-jobs which can
be utilized by all cookbooks with rakefiles (for now only includes a
job in the experimental queue).

Change-Id: Ie3687796faa210656d0b9d929fbc9e07de0d4a91
Blueprint: rakefile