Add a ci-sandbox project.
With the move to not using a special group for non voting CI accounts we
need a new sandbox project that allows all registered users to vote +/-1
in the Verified column.

The reason we did not modify the existing sandbox project is we do not
want to give new users that do not intend on performing CI the ability
to +/-1 Verified as they will never have to do that with normal
projects. This second sandbox project ensures other users get a
consistent experience.

Change-Id: Id70b4c5a8910d6d35742ed5cee9a9520ed2e2310