Make run-tox a super set of run-unittests
It turns out that we want unittest like things even when not running
unittests. Specifically we need things like subunit logs when running
functional tests. We should remove run-unittests and move to a single
tox runner script wherever possible. This means jobs simpler and results
in fewer surprises.

This change is the first in a few to make this happen. Step one
(implemented here) is to make run-tox.sh cover the use of
run-unittest.sh so that we can replace run-unittests.sh with run-tox.sh.
Once this is done and on all of our slave images we will convert all
jobs that use run-unittests.sh to run-tox.sh. Finally we can delete
run-tox.sh.

At some point we should consider doing similar for pep8 jobs and doc
jobs.

Change-Id: I457182df6a17cd599e528217cd7583212164f346