Use job-template for publishing to ci.openstack.org
Merge the various publishing jobs into a job template. The only
difference between them was the target and that the system-config one
used zuul_swift_upload_console_log.

The template now uses zuul_swift_upload_console_log for all the jobs.

The jobs using the template are now named "{name}-ci-docs" and needed to
be added to the jjb project and renamed in the zuul layout file.

Change-Id: I8772508ec9da4ba4b7d3bc37f85d260f80e836bc