Rename gerrit-powered-agenda to yaml2ical
Last part of the split between code and meeting descriptions in what was
previously known as gerrit-powered-agenda. Rename the code repository to
match new module name (following https://review.openstack.org/136427),
before we can publish it to PyPI in a subsequent change and and finally
make use of it in openstack-infra/irc-meetings gate checks.

Change-Id: I0fa99587f3d6e45c2546faded3cd55d5b2b83690