print pip freeze even when flake8 fails
When 'tox -e pep8' fails we don't end up printing
the pip freeze output since pipefail just bails out.
We should capture pip freeze when flake8 passes
or fails. The pipe seems to be a vestige from a previous
attempt to save the output into a separate file, we
dont need that anymore. Save the exit code from
the tox command and use that when we exit at the end
of the script.

Change-Id: Ic4590b7a8260a4a652608b1aa8d02beb786143ea