Invert post build script status logic
We had the logic in the post build script conditions inverted. We
unfortunately need to say onsuccess: False and onfailure: False to make
this run on every job regardless of its build status.

This is because the plugin treats those options as IFFs with onsuccess
being evaluated first. So if onsuccess is set to True then the post
build script will run if and only if the build status was successful.
Similar story for onfailure. If we set them both to false then neither
success nor failure is checked and we should run the script on all
builds.

Change-Id: I70fd948e239e8a5007b0f5813c9d5757a1742abe