Clean up Oslo stable compat jobs
Rename the sahara-stable-compat-jobs to make it more a more template
for stable compatibility jobs that only look back one release, to be
used for new projects that should not be tested for compatibility two
releases back.

Use this template for the Oslo libs created during Juno, and for the
Sahara project.

Change-Id: I9dee12e6b91e62dd16c547b300eb90abbbbd1072