Fix release publication for server projects
Designate and Sahara erroneously used the publish-to-pypi template.
It was convenient to generate tarballs on pre-release and release
tags, but it also resulted in a server tarball upload to PyPi,
which is not something we do for server projects. This commit
removes the extra jobs before we push the kilo-1 milestone.

Barbican and Zaqar did not use the merge-release-tags job, while
they probably should. Aligning them will allow to let all integrated
and incubated server projects use the same job template (proposed in
a subsequent commit).

Change-Id: Id46e00a05b2c943b51469f948b1d152f09be35f2