Add test for validating certain gerrit ACL rules
This patch adds a check for Gerrit ACL configs. In this initial
version, the test being added checks that the config defines
at least one team that acts as a core/admin one.

The current check is intentially pretty loose to ensure some
flexibility.

Currently the check applies to stackforge and openstack-dev only.

Change-Id: I2b04f2d94130699d6fd1c16a9b00b9c81d1ba07d