Introduce project-specific stable-maint teams
As discussed in Paris and subsequently discussed on the mailing-list,
switch stable branch maintenance to project-specific teams under the
guidance and leadership of the stable-maint-core team.

All $PROJECT-stable-maint teams in Gerrit should be owned by
stable-maint-core which will vet the addition of members to make sure
they are aware of the Stable Branch policy.

Remarks:
- Sahara was already using their own team, proposed change will align
  them to use a specific team instead.
- Designate is still incubated and therefore stable-maint-core doesn't
  own their stable branches yet. Proposed change fixes ACL to reflect
  that.
- All projects still inherit openstack-stable-maint from the
  all-projects ACL, but this will be removed once the transition is
  over.

Reference:
http://lists.openstack.org/pipermail/openstack-dev/2014-November/050390.html

Change-Id: Id127d24e4428be8b450de277dc55df20349aab13