zuul's log path should always be a folder
zuul sets the path for where logs are stored. In all our cases this
is a folder containing the console.log and other assets. We should
denote this is a folder by leaving a trailing slash. This also means
that os-loganalyze can determine if it should load an index.html
faster by looking for a trailing slash.

Change-Id: Ia1de627c01ee43c461d25b184fb61cbfd91fb19d