Only guard yum update against epel if enabled
The --disablerepo flag for yum will fail if the repo being disabled is
not already enabled. While this is not a problem for HP cloud images
that already have epel enabled, it makes it hard for users to use the
script on non-HP cloud nodes. This commit detects whether epel is
enabled and uses the disable flag appropriately.

Change-Id: I26b6fbdc390b99b0e289ac0e35a5e616d5d3682a