Run acceptance tests without rake
The puppetlabs_spec_helper has a 'spec' rake task that runs a defined
set of unit tests in the spec/ directory of a module. It is not
intended to accept any arguments. See [1] for the defined set. By
design, it does not run any tests in the spec/acceptance/ directory.
For modules with only tests in the spec/acceptance directory and
nowhere else in spec/, passing the argument spec/acceptance causes the
unforeseen and undefined behavior where it will appear to correctly run
the acceptance tests. An example of this strange behavior is [2]. For
modules that have unit tests in the spec/ directory, where unit tests
are defined by what the 'spec' rake task is intended to run, running
`rake spec spec/acceptance` causes the rake task to only run the unit
tests and not the acceptance tests. An example run is [3]. This patch
runs the tests directly with the rspec command instead of using the
rake command to go through the puppetlabs_spec_helper spec test rake
tasks. For reference, the intended way to run beaker is documented [4].

It should be noted that if a module does not have acceptance tests to
run, the test will fail whether we use the rake task or just the rspec
command. This commit is not intended to correct that, but to cause
modules that do have acceptance tests to run those tests and not
duplicate the effort of running unit tests that are already covered in
separate job.

[1] https://github.com/puppetlabs/puppetlabs_spec_helper/blob/master/lib/puppetlabs_spec_helper/rake_tasks.rb#L10
[2] http://logs.openstack.org/86/126086/12/check/gate-puppet-storyboard-puppet-beaker-rspec/074322d/console.html.gz
[3] http://logs.openstack.org/20/102020/13/check/gate-puppet-nova-puppet-beaker-rspec/c2a7dce/console.html.gz
[4] https://github.com/puppetlabs/beaker/wiki/How-to-Write-a-Beaker-Test-for-a-Module

Change-Id: I9defedb1010dae14896584ea5f6bf723a59a6356