Introduce Oslo-specific stable-maint team
As discussed in Paris and subsequently discussed on the mailing-list,
we are switching stable branch maintenance to project-specific teams
under the guidance and leadership of the stable-maint-core team.

Some Oslo libraries follow a "compatible" release cycle and therefore
use stable branches -- the proposed change is to create a unique
oslo-stable-maint team to review such changes. Oslo libraries which
follow a free release cycle (taskflow, oslosphinx, pbr...) do not use
stable branches and therefore do not need their ACLs changed.

The oslo-stable-maint team should be owned by stable-maint-core which
will vet the addition of members to make sure they are aware of the
Stable Branch policy.

Change-Id: Ic89244ecbdb317501f86ee4b7347a281b13cae91