Make prerelease and release pipelines match pep440
Our old release and prerelease regexes matched versions that are not
completely valid under pep440. Some that are not valid would be
normalized by setuptools to valid versions resulting in git tags that do
not match sdist python versions. Others are completely invalid and will
result in setuptools versions parse errors which in turn leads to odd
sort order for versions.

Avoid all of this oddness and require pep440 valid versions.

Change-Id: I19fd12f80ea68babfd4ea0344c556308e2e21a05