Generate tarballs for neutron-*aas on tags
openstack/neutron-*aas git repository are the result of a split from
neutron. Tarballs will be published as part of the release process on
milestones and final releases, and attached to the same Launchpad page
as Neutron itself.

To that effect, we need to enable the -tarball jobs to run on
pre-release and release zuul pipes for all neutron-*aas projects.

Change-Id: If9ce653c61d8bdadeec29d368207013a29f9e010