openstack/requirements stable branch ACL update
We recently switched stable branches changes to be reviewed by specific
teams instead of a common team. The openstack/requirements repository
makes use of stable branches but was left out of the previous changes
(since it does not rely on a project-specific team).

This commit assigns requirements stable branch reviews to the new,
cross-project, stable-maint-core team.

Change-Id: Iecc635824fcde2d00bf42ff6ebac0cc57735843d