The ci-sandbox project should not require an ICLA
The openstack-dev/ci-sandbox project is meant to be a place for
third-party testing systems administrators to test their
implementations, which sometimes involves creating changes in Gerrit
to validate stream-events triggers. Many of these sysadmins are not
software contributors to OpenStack projects and are likely to not
have signed the ICLA or joined the OpenStack Foundation, and so this
test project should not require such things thereby reducing the
number of prerequisites for them.

Change-Id: Id67603c26114abd05873ffd6f80194a691f5b3c3