We want to run freese in venv not with tox target
Running the freeze command as posargs to tox targets is unreliable as
those commands may be interpreted as arguments to other commands rather
than as commands themselves. For example `tox -epy26 -- pbr freeze` is
likely pass 'pbr freeze' as test name filters to the py26 test runner.

Instead what we want is to run pbr/pip freeze directly out of the venv
in question.

Change-Id: Iefb8ae4e82dcd31306d03befdc9766f2c1291571