Run freeze command after running tox
In the run-docs.sh script we attempted to run the freezecmd out of the
tox venv dirs before running tox. This means there was no tox venv to
run the command out of and it failed. Move the execution of the
freezecmd to after the initial tox run so that the freeze command is
available in the created virtualenv.

Fixes-bug: 1405451
Change-Id: I9d33aa1a06c5351e11181790641635cfc1d10874