Import shade into openstack-infra
Shade is a library that contains all the various logic you need to use
OpenStack with the clouds Infra has encountered. It's currently being
used as the basis of the next generation of ansible modules and the
intent is to replace the direct python-*client calls in nodepool and
launch_node.

Change-Id: I14332c43cf12d46bd4472312e8edfa5760fc2f68