Install test-requirements for swiftclient functional tests
The functional test for python-swiftclient currently fails all the time because
an required package for testing (testrepository) is missing.

swift itself uses nosetest to start functional tests, but python-swiftclient
uses testr and thus requires the package testrepository to be installed.

This patch simply installs all requirements for the tests. I was able to
reproduce this error localy using a fresh virtualenv environment, and this
patch fixed it for me.

Change-Id: I5417f74d324b1915155e321c3a87138069aa35fa