split up pbr tests from requirements tests
pbr installation testing needs to know that all current package
definitions will actually work upon a pbr change, in the multiple ways
that packages need to be installed.

requirements testing needs to ensure that all the requirements work,
but don't need to check that 3 times on every tree. They also only
need to care that the global list is installable (not currently tested).

We should fork these tests and specialize the requirements test to
it's purpose. Especially as doing so should reduce the run time by
over 50%.

Depends-On: I37b1510f3639c332a13bd7fc892b66f71b99b594

Change-Id: Ia49d93031425ce70e7667972fdf9078143d0dd53