Install argparse and requests to swift venv
The zuul swift log uploader script needs argparse on python2.6 and
requests on all python versions to work properly. Install these two
packages to the virtualenv configured for this script at image build
time.

Note that `pip install argparse` on python >= 2.7 is not an error
it just reports that argparse is installed as part of the python
installation and moves on.

Change-Id: Ie79b59e1c687cce2b435c9704959b02911f94778