Better name checking in jenkins project checks
Previously the check for ' name: ' was too relaxed and could match lines
that validly set a value for a macro. Make the check for 'name:' much
more restrictive to only match top level names in projects. This was you
can have macros that accept overridden name values without running into
trouble with alphabetization.

Change-Id: I16299f53433cdd1378ae6665399e3deb8a68d381