Set proper ownership on tempest test repository
When the testrepository directory is created by calling the preseed
script the permissions are different from the rest of the tempest
repo. When the dir is rsync'd by devstack-gate it errors out with
permission denied on the .testrepository dir. This commit fixes this
by setting the ownership to be the same as the rest of the cached
repo.

Change-Id: Ie12f3f0d824401f2ed53a77f8f307ea0f178854e