Fix centos6 image updates on hpcloud
The "yum repolist" command was failing with what I think was the
same error that the update command was designed to fix.  I am
unsure how or if that ever worked correctly, or what might have
changed to cause it to start impacting that command.

At any rate, specifying "epel*" as the repo name to disable will
cause the update command to work regardless of whether it is
installed.

Change-Id: I1d978ade76ec78e8895e00bbf64e5e40529223f8