Upload console and logs together for index gen
We get one shot at uploading a set of correct indexes to swift. If we do
multiple passes at uploading swift logs then we can have incomplete
indexes as the index generator will only have a partial view of the
world.

Switch the dg swift upload publisher to upload console.html and
the devstack logs together so that we generate a complete and correct
root index that includes the console log and the logs/ dir.

Change-Id: I6e9b6184deaea9f9aa43bc329147f2e9a9622306