Remove stable-compat-jobs from Oslo libraries
We are moving to cap all of our libraries in stable branches, so there
is no longer a need to test changes to master under the stable
branches. This change removes the jobs from Oslo libraries. Although the
caps are not yet in place, we are holding off on making any more
releases until the current stable/juno requirements issue is resolved
and the caps are in place. In the mean time, removing these jobs
unblocks work on the master branches of the libraries.

Change-Id: Id8e0381beb68efa6a8cd0979cc5e85df2541e8a0