Add extra information to zuul swift uploading
To help see if generating or uploading is slow with zuul-swift,
add in some crude debugging. This will never be displayed as the
console log is fetched before uploading is done so to see the
debug messages you'll have to look at the logs Jenkins captures.

The useful information will be to look at the differences in the
timestamps on the messages captured by Jenkins.

Change-Id: I7c31084a92218c3d12ebed68717088337b06f905