skip rerunning requirements changes in devstack
Requirements changes to projects trigger tempest/grenade
jobs. However, these are exactly the same tests as would have just
been run in the requirements job before we propose requirements
changes to projects. So the merge of a requirements change current
creates a great thundering herd in all the projects.

Skip if this is the proposal bot and the only content changed is
requirements. The requirements check job will ensure that people can't
sneak an invalid requirement range in.

Change-Id: Ie0c7470eae497ec666f200d2119f5de415901d1b